agent-service/pyproject.toml
[project]
name = "agent-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "chromadb>=0.6.3",
    "fastapi[standard]>=0.115.12",
    "jose>=1.0.0",
    "langchain>=0.3.24",
    "langchain-community>=0.3.23",
    "langchain-openai>=0.3.16",
    "langgraph>=0.4.1",
    "langsmith>=0.3.39",
    "pydantic-settings>=2.9.1",
    "pymilvus>=2.5.8",
    "python-jose>=3.4.0",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "ruff>=0.11.9",
]
